requirements.txt
tensorflow
keras
scipy
scikit-learn
opencv-python
matplotlib
imutils
argparse
pillow
streamlit